Update secret strategy, add secret to deployment
Only put known required configuration values in the secret created
when deploying the SDM.  Add use of the secret to the deployment
spec.  I am still not sure whether only doing this in local mode is
the right decision.

Set application namespace to "sdm".

Annotate ingress spec for nginx-ingress if in local mode and
Kubernetes configuration context is "minikube", but only if an ingress
would otherwise be created.

Move Docker config under build, as per the custom in sdm-pack-docker.

Refactor out getting current Kubernetes configuration context into a
function so it can be used a couple places.